Cache the readers, now list_readers returns a tuple
We need to cache the readers because when connecting to a card we must
use the name, not an index or identifier, however the application will
use zero-based indexing (e.g. `acr show --card 0`).

It will be much easier to implement if we cache the readers and the
readers count.

Also now `list_readers` returns a `tuple` instead of a `list`, always
the same object.